+++ swri_geometry_util/include/swri_geometry_util/geometry_util.h
@@ -30,6 +30,7 @@
 #ifndef GEOMETRY_UTIL_GEOMETRY_UTIL_H_
 #define GEOMETRY_UTIL_GEOMETRY_UTIL_H_
 
+#include <opencv2/core/core.hpp>
 #include <tf/transform_datatypes.h>
 
 namespace swri_geometry_util
@@ -58,10 +59,52 @@ namespace swri_geometry_util
       const tf::Vector3& line_end,
       const tf::Vector3& point);
 
+  double DistanceFromLineSegment(
+      const cv::Vec2d& line_start,
+      const cv::Vec2d& line_end,
+      const cv::Vec2d& point);
+
   tf::Vector3 ProjectToLineSegment(
       const tf::Vector3& line_start,
       const tf::Vector3& line_end,
       const tf::Vector3& point);
+
+  cv::Vec2d ProjectToLineSegment(
+      const cv::Vec2d& line_start,
+      const cv::Vec2d& line_end,
+      const cv::Vec2d& point);
+
+  /**
+   * Test if a point is within the bounds of a polygon.
+   *
+   * The polygon is assumed to be non-intersecting and that the vertices
+   * are ordered sequentially in clock-wise fashion.
+   *
+   * @param[in]  polygon  The list of polygon vertices.
+   * @param[in]  point    The point.
+   *
+   * @return True if the point is within the bounds of the polygon.
+   */
+  bool PointInPolygon(
+      const std::vector<cv::Vec2d>& polygon,
+      const cv::Vec2d& point);
+
+  /**
+   * Calcuate the distance between a point and the bounds of a polygon.
+   *
+   * The polygon vertices are assumed to be ordered sequentially.
+   *
+   * The distance returned is positive even if the point is within the polygon.
+   *
+   * @param[in]  polygon  The list of polygon vertices.
+   * @param[in]  point    The point.
+   *
+   * @return The distance from the point to the polygon.  -1 is returned if the
+   *         polygon is empty.
+   */
+  double DistanceFromPolygon(
+      const std::vector<cv::Vec2d>& polygon,
+      const cv::Vec2d& point);
 
   /**
    * Find closest point to two 3D lines.